Start of attempt to break up composite slots of ResponsiveTable into reusable composite components with slot forwarding to make customizing parts easier - only to discover slot forwarding is BROKE AF in svelte.
Also wanted something that would format the lowlevel recursive types less as @html strings and more as properly manipulable svelte patterns. Leaving format in unsatisfactory state for now but available for slot forwarding.

Going forward at this point with folder for related components and documentation in place and focusing on creating a FieldMultiple like implementation with ResponsiveTable (as is) as the layout of the fields.